+++ src/App.jsx
@@ -5,18 +5,11 @@ import MovieCard from './components/MovieCard';
 import 'swiper/css';
 import { Navigation, Pagination } from 'swiper/modules';
 import { Swiper, SwiperSlide } from 'swiper/react';
+import { API_URL } from './data/movieApi';
 import { useFetch } from './hook/useFetch';
-//TODOS data:RTK로 관리하기
-
-// export const BASE_IMG_URL = import.meta.env.VITE_BASE_IMG_URL;
-export const BASE_IMG_URL = 'https://image.tmdb.org/t/p/w500';
-//TMDB 사용
-export const API_KEY = import.meta.env.VITE_MOVIE_API_KEY;
-const API_URL = 'https://api.themoviedb.org/3/movie/popular?language=ko&page=1';
 
 export function App() {
   const movies = useFetch(API_URL);
-  console.log(movies.results);
   return (
     <>
       <h1 className="bg-amber-300 p-3 text-center text-3xl">이달의 영화</h1>
@@ -31,7 +24,7 @@ export function App() {
       >
         {movies.results &&
           movies.results.map((movie) => (
-            <SwiperSlide key={movie.id} className="border-2 px-4">
+            <SwiperSlide key={movie.id} className="mb-5 px-4">
               <Link to={`/details/${movie.id}`}>
                 <MovieCard movie={movie}></MovieCard>
               </Link>
