+++ src/components/MovieDetail.jsx
@@ -1,4 +1,4 @@
-import { BASE_IMG_URL } from '@/App';
+import { BASE_IMG_URL, DETAIL_URL } from '@/data/movieApi';
 import { useFetch } from '@/hook/useFetch';
 import { useParams } from 'react-router-dom';
 
@@ -6,12 +6,9 @@ export default function MovieDetail() {
   //객체로 id값 받아오기
   const { id } = useParams();
   console.log(id);
-  const detail = useFetch(
-    `https://api.themoviedb.org/3/movie/${id}?language=ko`,
-  );
-  console.log(detail);
+  const detail = useFetch(`${DETAIL_URL}${id}?language=ko`);
   return (
-    <div className="mx-auto grid max-w-[1080px] grid-cols-8 grid-rows-5 gap-3 border-2 p-5">
+    <div className="mx-auto grid max-w-[1080px] grid-cols-8 grid-rows-5 gap-3 p-5">
       <img
         src={`${BASE_IMG_URL}${detail.backdrop_path}`}
         alt="posterImg"
@@ -27,7 +24,7 @@ export default function MovieDetail() {
       </h3>
       <div className="col-span-4 row-span-1 flex items-center justify-around bg-indigo-300 p-3">
         장르{' '}
-        {detail?.genres?.map((itm) => (
+        {detail.genres?.map((itm) => (
           <p key={itm.id} className="rounded-3xl border px-3">
             {itm.name}
           </p>
